gcloud-skyscanner/src/CF-Skyscanner-ProcessRow/requirements.txt
# Function dependencies, for example:
# package>=version
requests
google-cloud-bigquery
google-cloud-pubsub
datetime